serverApp/requirements.txt
asgiref==3.4.1
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
pip==23.0.1
pycparser==2.21
PyJWT==2.7.0
pytz==2022.7
setuptools==66.0.0
six==1.16.0
sqlparse==0.4.2
trollius==2.1.post2
typing_extensions==4.5.0
wheel==0.38.4
pusher==3.3.2
python-dotenv==1.0.1
